requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
flask-wtf
python-dotenv
reportlab
werkzeug
selenium
python-dotenv
pytest
pytest-html